+++ drivers/media/platform/exynos/mfc/s5p_mfc_common.h
@@ -19,9 +19,9 @@
 #include <linux/workqueue.h>
 #include <linux/slab.h>
 #include <linux/videodev2_exynos_media.h>
+#include <linux/exynos_mfc_media.h>
 
 #include "exynos-mfc.h"
-#include "exynos_mfc_media.h"
 
 #if defined(CONFIG_EXYNOS_MFC_V10)
 #include "regs-mfc-v10.h"
